+++ Sensors/SMCSuperIO/SMCSuperIO.cpp
@@ -141,27 +141,32 @@ IOReturn SMCSuperIO::setPowerState(unsigned long state, IOService *whatDevice) {
 }
 
 SuperIODevice* SMCSuperIO::detectDevice() {
-	if (PE_parse_boot_argn("ssioec", &deviceNameEC, sizeof(deviceNameEC))) {
-		DBGLOG("ssio", "found EC device %s", deviceNameEC);
-	} else {
-		auto name = getParentEntry(gIOServicePlane)->getProperty("ec-device");
-		auto nameStr = OSDynamicCast(OSString, name);
-		auto nameData = OSDynamicCast(OSData, name);
-		if (nameStr) {
-			strlcpy(deviceNameEC, nameStr->getCStringNoCopy(), sizeof(deviceNameEC));
-			DBGLOG("ssio", "found EC device %s from string", deviceNameEC);
-		} else if (nameData) {
-			auto s = nameData->getLength();
-			if (s > sizeof(deviceNameEC)) s = sizeof(deviceNameEC);
-			strlcpy(deviceNameEC, static_cast<const char *>(nameData->getBytesNoCopy()), s);
-			DBGLOG("ssio", "found EC device %s from data", deviceNameEC);
+	SuperIODevice* detectedDevice;
+
+	auto lpc = getParentEntry(gIOServicePlane);
+	if (lpc != nullptr) {
+		if (PE_parse_boot_argn("ssioec", &deviceNameEC, sizeof(deviceNameEC))) {
+			DBGLOG("ssio", "found EC device %s", deviceNameEC);
+		} else {
+			auto name = lpc->getProperty("ec-device");
+			auto nameStr = OSDynamicCast(OSString, name);
+			auto nameData = OSDynamicCast(OSData, name);
+			if (nameStr) {
+				strlcpy(deviceNameEC, nameStr->getCStringNoCopy(), sizeof(deviceNameEC));
+				DBGLOG("ssio", "found EC device %s from string", deviceNameEC);
+			} else if (nameData) {
+				auto s = nameData->getLength();
+				if (s > sizeof(deviceNameEC)) s = sizeof(deviceNameEC);
+				strlcpy(deviceNameEC, static_cast<const char *>(nameData->getBytesNoCopy()), s);
+				DBGLOG("ssio", "found EC device %s from data", deviceNameEC);
+			}
+		}
+		detectedDevice = EC::ECDevice::detect(this, deviceNameEC, lpc);
+		if (detectedDevice) {
+			return detectedDevice;
 		}
 	}
 
-	SuperIODevice* detectedDevice = EC::ECDevice::detect(this, deviceNameEC);
-	if (detectedDevice) {
-		return detectedDevice;
-	}
 	detectedDevice = WindbondFamilyDevice::detect(this);
 	if (detectedDevice) {
 		return detectedDevice;